refactor(web): Bearer トークンを統一し X-Session-Token を廃止
admin JWT と anonymous JWT を共に Authorization: Bearer で送信する方式に統一。
getBearerToken() で admin JWT を優先、なければ session JWT を使用。
session-token.ts を auth-token.ts に統合し、X-Session-Token ヘッダーを廃止。
